Use prettier jasmine test reporter to make it easier work with tests
