+++ package.json
@@ -187,6 +187,7 @@
 		"@types/vscode": "^1.93.0",
 		"@vscode/test-electron": "^2.4.1",
 		"jasmine": "^5.3.1",
+		"jasmine-console-reporter": "^3.1.0",
 		"jasmine-core": "^5.3.0",
 		"minimist": "^1.2.8",
 		"tslint": "^6.1.3",
